Extract `MaybeDangling` to `util` mod
