+++ tokio-util/src/sync/cancellation_token.rs
@@ -4,6 +4,7 @@ pub(crate) mod guard;
 mod tree_node;
 
 use crate::loom::sync::Arc;
+use crate::util::MaybeDangling;
 use core::future::Future;
 use core::pin::Pin;
 use core::task::{Context, Poll};
@@ -93,46 +94,8 @@ pin_project! {
         // See <https://users.rust-lang.org/t/unsafe-code-review-semi-owning-weak-rwlock-t-guard/95706>
         // for more info.
         #[pin]
-        future: MaybeDanglingFuture<tokio::sync::futures::Notified<'static>>,
+        future: MaybeDangling<tokio::sync::futures::Notified<'static>>,
         cancellation_token: CancellationToken,
-    }
-}
-
-/// A wrapper type that avoids implicitly asserting the contents are currently valid.
-/// It's safe to contain a dangling reference, as long as it is not used.
-///
-/// # Safety
-///
-/// The inner type must be dropped before the data it references to.
-///
-// Reference
-// See <https://users.rust-lang.org/t/unsafe-code-review-semi-owning-weak-rwlock-t-guard/95706>
-//
-// TODO: replace this with an official solution once RFC #3336 or similar is available.
-// <https://github.com/rust-lang/rfcs/pull/3336>
-#[repr(transparent)]
-struct MaybeDanglingFuture<T>(core::mem::MaybeUninit<T>);
-
-impl<T> Drop for MaybeDanglingFuture<T> {
-    fn drop(&mut self) {
-        // Safety: `0` is always initialized.
-        unsafe { core::ptr::drop_in_place(self.0.as_mut_ptr()) };
-    }
-}
-
-impl<T> MaybeDanglingFuture<T> {
-    fn new(inner: T) -> Self {
-        Self(core::mem::MaybeUninit::new(inner))
-    }
-}
-
-impl<F: Future> Future for MaybeDanglingFuture<F> {
-    type Output = F::Output;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        // Safety: `0` is always initialized.
-        let fut = unsafe { self.map_unchecked_mut(|this| this.0.assume_init_mut()) };
-        fut.poll(cx)
     }
 }
 
@@ -329,7 +292,7 @@ impl WaitForCancellationFutureOwned {
             // # Safety
             //
             // cancellation_token is dropped after future due to the field ordering.
-            future: MaybeDanglingFuture::new(unsafe { Self::new_future(&cancellation_token) }),
+            future: MaybeDangling::new(unsafe { Self::new_future(&cancellation_token) }),
             cancellation_token,
         }
     }
@@ -370,7 +333,7 @@ impl Future for WaitForCancellationFutureOwned {
             // # Safety
             //
             // cancellation_token is dropped after future due to the field ordering.
-            this.future.set(MaybeDanglingFuture::new(unsafe {
+            this.future.set(MaybeDangling::new(unsafe {
                 Self::new_future(this.cancellation_token)
             }));
         }
